#96 [Rename] Controller 오타 수정
